requirements.txt
openpyxl
pdfplumber
pytesseract
python-dotenv
flask
openai
pdf2image